Add the start of a kernel driver tracer, panwrap
This adds the start of a libwrap equivalent for panloader, called
panwrap. This library can be preloaded using LD_PRELOAD for any GL
application on android, and will log all of the ioctls made by the
respective application along with decoding it into human readable form
to the best of our ability.

Right now we only have decoding functions for the GET_VERSION ioctl and
the SET_FLAGS ioctl, but more are on the way!