panwrap: Decode the last ioctls we need for uapi tracing
This adds very basic decoding for the last ioctl's we need to look at in
order to look into how to make this work the uapi of the current kernel
driver.

As well, this adds PAD_PTR which does the weird pointer padding thing
that the kernel driver expects, except without using the crappy typedef
provided to us by the current kernel drivers. I haven't actually tested
this at all with 32 bit arm systems, but right now we're not too
concerned about making those work. If you ever have any problems with
that though, this commit is the reason.