panwrap: Do exact flag comparison when decoding flags
For certain things, mali made the weird decision of making flags that
technically contain other flags. For instance:

 #define BASE_JD_REQ_SOFT_DUMP_CPU_GPU_TIME      (BASE_JD_REQ_SOFT_JOB | 0x1)
 #define BASE_JD_REQ_SOFT_FENCE_TRIGGER          (BASE_JD_REQ_SOFT_JOB | 0x2)
 #define BASE_JD_REQ_SOFT_FENCE_WAIT             (BASE_JD_REQ_SOFT_JOB | 0x3)

FENCE_WAIT contains both of the previous flags. The only way we can
actually decode this properly is to make sure that we don't just check
that some of the bits are set that match the flag we're checking, but
all of them.