panwrap: Add initial job/sync hexdumping
Now that we can track all of the memory maps that the bifrost GL driver
creates with the kernel driver, we actually know the size of all of the
memory regions being used to communicate with the kernel driver and as
such can start logging hexdumps of data from those regions.

As well some of these hexdumps can be a bit large, and some dumps are
very likely to have huge sections of uninitialized data we don't care
about. For those, we actually trim the size of the hexdump so as to only
contain the initialized memory in the memory region we're dumping. When
these hexdumps are trimmed, we also politely make note of it in
panwrap's output.