panwrap: Make timestamps more accurate by freezing time
The timestamps feature I added didn't really help me figure out much of
anything last time, but after trying to use it I realized that we should
try not to include any of the time spent in panwrap's logging functions
in the timestamps, so to make the offsets between system calls more
accurate since certain logging functions and memory tracking can become
expensive.

To workaround this and increase the overall accuracy of timestamps, we
make it look like we're able to stop the monotonic timer while logging
things. In actuality, we're just recording the time before pretending to
stop the timer, then recording the time when we pretend to restart the
timer and adding it to a timestamp value that solely consists of how
long we've spent in panwrap functions. Then when getting the timestamp,
we use the monotonic clock with the time we've spent in panwrap
functions subtracted.

Hopefully this silliness ends up actually being useful.