panwrap: Add decoding of actual job types
Looking at the job submission in the kernel driver it appears that the
core_req flags are a bit more complex then I expected, and also that we
can actually discern the type of job beyond just it's core requirements
based off of that data. Now we can see where BOTH vertex shaders and
fragment shaders start!