include/mali-ioctl.h: Define new type mali_gpu_ptr
This is just a simple typedef which represents the correct size of a GPU
pointer depending on the arch that panwrap is compiled for. This should
be used over a plain void*, since (while this seems to be the case on my
Hikey960, you can never be too sure) there's no guarantee that GPU VAs
will always be dereferenceable. Additionally, we also add some macros to
use with printf.