panwrap: Add 32bit size checks for ASSERT_SIZEOF_TYPE
At the moment, all of the structs for the ioctl arguments we're working
with are ensured to be the same size on 32bit and 64bit, however this
isn't going to be the case for some of the actual compiled shader
structs. So we need to be able to record different struct sizes for each
arch if necessary.