panwrap: Add the start of a job chain decoder, yay!
This is completely based off the code that was originally in pantry,
which for the time being can be considered abandoned. Some parts have
been copy pasted, but most of this has been completely rewritten for a
couple of reasons:

- The old code was really kind of difficult to understand, a lot of hard
  32-bit assumptions were made along with lots of sparse incomplete bits
  of code here and there
- The old code never worked on my 960 anyway :(
- Help me get a better understanding of what this code actually does and
  what we know so far collectively

Expect some things to be completely wrong, and I suggest you use the
original chai/pantry implementations as reference for "correct"
behavior, at least for older GPUs.

Additionally: I suspect a lot of these fields might be completely wrong.
But we have to start somewhere I suppose!