panwrap: Improve fbd logging, make it louder
A lot of the fields we thought we had decoded thus far in the fbd appear
to be incorrect, or at least they're not correct for all different job
types. I'm pretty sure I've corrected any incorrect looking offsets I
could find in the mali_tentative_mfbd struct but there's definitely
more. So:

- Print the offset of every single field we decode, to make referencing
  against the undecoded hexdump and noticing incorrect offsets easier
- Turn the raw fbd pointer (e.g. with flags, fbd type, etc.) into a
  struct with appropriate compiletime size checking, use it in the
  vertex/tiler payload (we'll reuse this struct in the next commit)
- Check that the fbd pointer is actually non-zero, and don't dump it
  otherwise (fixes a lot of invalid gpu memory access aborts)