panwrap: Import decoding for fragment payloads in the job chain
This imports the work that cafe did with chai's tracer to trace the
fragment payloads in job chains. There isn't anywhere as much to trace
here, although we do see ourselves reusing the fbd_meta struct from the
last commit as they appear to be important for fragment payloads as
well.

Pretty sure most of the information here is correct, with the only
unknown parts being the supposed flags embedded in the tile coords.
Additionally, we don't use bitfields for this struct since we don't know
which part of the tile coords actually contain flags yet, and doing so
pre-emptively might make things more difficult to work with.